# docs/swagger.yaml
definitions:
...
  group.GroupInfo:
    properties:
      createTime:
        type: integer
      faceUrl:
        type: string
      groupId:
        type: string
      groupName:
        type: string
      introduction:
        type: string
      memberCount:
        type: integer
      notification:
        type: string
      ownerId:
        type: string
    type: object
...
  group.paramsGetGroupInfo:
    properties:
      groupIDList:
        items:
          type: string
        type: array
      operationID:
        type: string
    required:
    - groupIDList
    - operationID
    type: object
...
  user.result:
    properties:
      data: {}
      errCode:
        example: 0
        type: integer
      errMsg:
        example: error
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /group/get_groups_info:
    post:
      consumes:
      - application/json
      description: get groups info
      parameters:
      - description: get groups info params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.paramsGetGroupInfo'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user.result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/group.GroupInfo'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - group
...
swagger: "2.0"